Update Directory structure
