+++ .gitignore
@@ -1,15 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-/build
+/client/node_modules
+/server/node_modules
 
 # misc
 .DS_Store
