+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "jango-ecommerce-store",
+  "name": "jango-ecommerce-store-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -34,5 +34,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "tailwindcss": "^3.4.1"
   }
 }
