+++ client/src/App.js
@@ -1,0 +1,7 @@
+const App = () => {
+  return (
+    <h3 className="text-3xl text-center text-bold">Jango Ecommerce Store</h3>
+  );
+}
+
+export default App;
