+++ server/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "jango-ecommerce-store-server",
+  "version": "1.0.0",
+  "private": true,
+  "main": "server.js",
+  "scripts": {},
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.18.2"
+  },
+  "devDependencies": {
+		"nodemon": "^3.1.0"
+	}
+}
