+++ server/server.js
@@ -1,0 +1,1 @@
+console.log("Server")
