--- src/App.js
@@ -1,7 +1,0 @@
-const App = () => {
-  return (
-    <h3>Jango Ecommerce Store</h3>
-  );
-}
-
-export default App;
